08.03.2021

+++ accounts/views.py
@@ -1,15 +1,23 @@
-from django.shortcuts import render
+# region import libraries
 from django.http.response import JsonResponse
-from .serializer import AccountsSerializer
-from django.contrib.auth.models import User
-from rest_framework.viewsets import ViewSet
-from rest_framework.decorators import action
-from rest_framework_jwt.settings import api_settings
-from rest_framework import permissions
-from .serializer import AccountsSerializer
-from knox.views import LoginView
-from knox.models import AuthToken
 from django.contrib.auth import authenticate
+from django.contrib.auth.models import User
+from django.shortcuts import render
+
+from .serializer import AccountsSerializer
+from .serializer import AccountsSerializer
+
+from knox.models import AuthToken
+from knox.views import LoginView
+
+from rest_framework.serializers import ValidationError
+from rest_framework_jwt.settings import api_settings
+from rest_framework.response import Response
+from rest_framework.decorators import action
+from rest_framework.viewsets import ViewSet
+from rest_framework import permissions
+
+# endregion
 
 
 class AccountsViewSet(ViewSet):
@@ -29,6 +37,24 @@ class AccountsViewSet(ViewSet):
         }
         return JsonResponse(response)
 
+    def list(self, request, *args, **kwargs):
+        instance = request._auth and request._auth.user
+        if instance is None:
+            raise ValidationError("Invalid token")
+        serializer = AccountsSerializer(instance)
+        return Response(serializer.data)
+
+    def retrieve(self, request, *args, **kwargs):
+        return JsonResponse({"message": "Bad request"})
+
+    @action(
+        detail=False,
+        methods=["GET"],
+        permission_classes=[permissions.IsAuthenticated],
+    )
+    def check_token(self, request, *args, **kwargs):
+        return JsonResponse({"token": True})
+
     @action(detail=False, methods=["POST"])
     def login(self, request, *args, **kwargs):
         username = request.data.get("username", None)
@@ -41,34 +67,20 @@ class AccountsViewSet(ViewSet):
         _, token = AuthToken.objects.create(user)
         return JsonResponse({"token": token})
 
-    @action(
-        detail=False,
-        methods=["POST"],
-        permission_classes=[permissions.IsAuthenticated],
-    )
-    def logout(self, request, *args, **kwargs):
+    def update(self, request, *args, **kwargs):
+        instance = request._auth and request._auth.user
+        if instance is None:
+            raise ValidationError("Invalid token")
+        serializer = AccountsSerializer(
+            instance=instance, data=request.data, partial=True
+        )
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(serializer.data)
+
+    def destroy(self, request, *args, **kwargs):
+        instance = request._auth and request._auth.user
+        if instance is None:
+            raise ValidationError("Invalid token")
         request._auth.delete()
         return JsonResponse({"message": "logout"})
-
-    @action(
-        detail=False,
-        methods=["GET"],
-        permission_classes=[permissions.IsAuthenticated],
-    )
-    def check_token(self, request, *args, **kwargs):
-        return JsonResponse({"token": True})
-
-    @action(
-        detail=False,
-        methods=["GET"],
-        permission_classes=[permissions.IsAuthenticated],
-    )
-    def get_user(self, request, *args, **kwargs):
-        user = request._auth.user
-        response = {
-            "first_name": user.first_name,
-            "last_name": user.last_name,
-            "username": user.username,
-            "email": user.email,
-        }
-        return JsonResponse(response)
